+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.4.0] 2025-10-05
+
+### Added
+
+- ETL_engineering: 
+    - `TemperatureExtractor`, extracts temperature values from a string column.
+    - `MultiTemperatureExtractor`, extracts multiple temperature values from a single string column into several new columns.
+    - `AutoDummifier`, now has an optional 'drop_first' parameter.
+
 ## [10.3.0] 2025-10-05
 
 ### Added